refactor(views): modularize views_gcal + views_gcal_dashboard into views_gcal/ package
Consolidates views_gcal.py (94 lines) and views_gcal_dashboard.py (1364 lines)
into a modular package structure with 7 files organized by domain:

Structure:
- views_gcal/__init__.py - Re-exports (18 exports)
- views_gcal/gcal.py - Core GCal (gcal_calendars, gcal_health)
- views_gcal/helpers.py - Helpers (_filter_events_queryset, _apply_common_filters, DashboardEventsPagination)
- views_gcal/summary.py - 4 views (GCalStatusSummaryView, GCalListView, DashboardMetricsView, AlertsSummaryView)
- views_gcal/batch.py - 3 views (GCalPublishBatchView, GCalBatchReapplyView, GCalBatchResyncView)
- views_gcal/detail.py - 4 views (DashboardEventsView, DashboardEventsExportView, EventDetailAPIView, GCalDriftView)
- views_gcal/insights.py - 2 views (SuccessRateView, TopInsightsView)

Changes:
- Delete views_gcal.py (94 lines)
- Delete views_gcal_dashboard.py (1364 lines)
- Create views_gcal/ package with 7 files
- Update urls.py imports
- Add backwards compatibility test (20 tests)

Tests: 809 passed, 22 skipped
